requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pytest>=8.0.0
httpx>=0.27.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
yars>=1.2.0
notion-client>=2.2.1
fastapi>=0.115.0
uvicorn>=0.30.0
